MeshManager doesn't cause SegFault produce cube with appropriate normals
each face has normals point in its direction

+++ main.cpp
@@ -124,7 +124,7 @@ void setView() {
 
 void getInputState()
 {
-   for (int i = 0; i < kart_objects.size(); i++) {
+   for (int i = 0; i < (int)kart_objects.size(); i++) {
       float joy[4]; //should vary from -1.0 to 1.0
       unsigned char button[32]; //either GLFW_PRESSED or GLFW_RELEASED
